experience/6309: make ValidationReceiver it's own class instead of subclass of SubmissionReceiver
