+++ prime-router/src/main/kotlin/azure/ValidateFunction.kt
@@ -1,5 +1,6 @@
 package gov.cdc.prime.router.azure
 
+import ValidationReceiver
 import com.google.common.net.HttpHeaders
 import com.microsoft.azure.functions.HttpMethod
 import com.microsoft.azure.functions.HttpRequestMessage
@@ -15,7 +16,6 @@ import gov.cdc.prime.router.ActionLogLevel
 import gov.cdc.prime.router.InvalidReportMessage
 import gov.cdc.prime.router.Sender
 import gov.cdc.prime.router.Translator
-import gov.cdc.prime.router.ValidationReceiver
 import gov.cdc.prime.router.azure.db.enums.TaskAction
 import gov.cdc.prime.router.common.JacksonMapperUtilities
 import gov.cdc.prime.router.history.Destination
